LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/navigation-preload/resource-timing.https-expected.txt



FAIL Navigation Preload Resource Timing. assert_greater_than_equal: domainLookupStart >= requestStart of https://web-platform.test:9443/service-workers/service-worker/navigation-preload/resources/resource-timing-scope.py?type=redirect expected a number greater than or equal to 17 but got 0
